remove unnecessary function

+++ CPP06/ex01/Serializer.cpp
@@ -1,9 +1,9 @@
 #include "includes.hpp"
 
-unsigned long Serializer::serialize(Data* ptr) {
-    return reinterpret_cast<unsigned long>(ptr);
+uintptr_t Serializer::serialize(Data* ptr) {
+    return reinterpret_cast<uintptr_t>(ptr);
 }
 
-Data* Serializer::deserialize(unsigned long raw) {
+Data* Serializer::deserialize(uintptr_t raw) {
     return reinterpret_cast<Data*>(raw);
 }
