+++ CPP06/ex01/Serializer.hpp
@@ -13,7 +13,7 @@ private:
     Serializer& operator=(const Serializer& copy);
 
 public:
-    static unsigned long serialize(Data* ptr);
+    static uintptr_t serialize(Data* ptr);
     static Data* deserialize(unsigned long raw);
 };
